[Issue 963] Moves ECS services into private subnets (#1014)
## Summary

Fixes https://github.com/HHS/simpler-grants-gov/issues/963

This PR will be split into 3 major stages:

### Stage 1: Create private subnets

The existing networking infrastructure is in its "default" (AWS provided) state. This state does not meet the standards of our [ideal network layer design](https://github.com/navapbc/template-infra/blob/main/docs/decisions/infra/0011-network-layer-design.md) in a few ways. But the most notable one is: the default networking infrastructure lacks private subnets. You cannot remove the public IP addresses from the ECS instances without a first creating private subnets. So I'm requesting code review for Stage 1, backfilling the private subnets into the default VPC. This also includes attaching the VPC endpoints onto those private subnets.

Deploying this will require me to deploy the subnets from my branch, before I merge. I've talked about this with @aplybeah. Essentially, I need to (temporarily) ignore "only deploy from main" rule in order to resolve https://github.com/HHS/simpler-grants-gov/issues/963 in a timely manner. The alternative would be for me to take several days to deploy an entirely new networking layer from scratch, which feels like too much work for this task / our timeline.

Due to the above, there will be a long running (eg. for a few days) terraform state diff on the `infra/networks/` folder.

### Stage 2: Creating a "staging" site inside the private subnets

When I create the private subnets, I won't be able to be sure that they are totally working for the sake of deploying our application. There's a chance this could be a breaking change from the point of view of the application. So it requires testing, specifically testing in the form of creating a 3rd simpler ECS application. I'm calling this application "staging", since "dev" and "prod" are already in use. I'll deploy this application to the private subnets to confirm that the application is working. I'll then keep the application around so that we can use it for other purposes later.

### Stage 3: Move all sites into the private subnets

After confirming that staging works, I'll make sure the terraform configuration for dev and prod are using the new subnets as well. I'll deploy to dev and prod after I merge. This will not change the load balancer URL, but it might cause a few minutes of downtime. I expect that any potential downtime will be 5 minutes at most, enough time for the new app containers to start up.

### Testing

Deploy the networking layer changes:

```
$ cd infra/networks/
$ terraform apply
```

Deploy the frontend application (with the latest build as of Jan 16th)

```
$ cd infra/frontend/service
$ terraform init -backend-config=staging.s3.tfbackend -reconfigure
$ terraform apply \
    -var="environment_name=staging" \
    -var="image_tag=2698a01eca3a7373ae4ddeec78271108f50234fd"
```

Deploy the database

```
$ cd infra/api/database
$ terraform init -backend-config=staging.s3.tfbackend -reconfigure
$ terraform apply \
    -var="environment_name=staging"
```

Deploy the backend application (with the latest build as of Jan 16th)

```
$ cd infra/api/service
$ terraform init -backend-config=staging.s3.tfbackend -reconfigure
$ terraform apply \
    -var="environment_name=staging" \
    -var="image_tag=2698a01eca3a7373ae4ddeec78271108f50234fd"
```

### Future Work

In the future, I would like to move to the network architecture described [in the Nava infra template](https://github.com/navapbc/template-infra/blob/main/docs/decisions/infra/0011-network-layer-design.md). That design specifies a separate network for every environment. This would be an improvement over our current design, where there is one network for all environments. The fact that there's only 1 network is what makes testing stage 1 so complicated.

This PR includes a new `staging` app, but does not (yet) add that staging app to Github Actions. I intend to do that at a later date!