[Issue 1085] Make the topportunity transfer table and connect the API directly to it (#1103)
## Summary
Fixes #1085

### Time to review: __5 mins__

## Changes proposed
Added a `transfer_topportunity` table for us to copy directly to from
the current Oracle database

Temporarily modified the API to use this table directly (has a few
conversions to make it generally the same as before) - this will be
removed/replaced when we setup the transformation layer in the next
phase of the API.

Updated ERD generation to include transfer tables

Updated factories and local seeding script to generate data in the
transfer table as well

## Context for reviewers
We need tables to copy from the existing Oracle DB into our Postgres DB
as transfer tables. While we need to do some amount of testing of DMS,
for simplicity we'll try to keep the tables as _close_ to their existing
Oracle structure as possible, with a few minor caveats. The actual DB
models were generated from a script that reads the existing Oracle
schema and outputs the model (can include, although it's deliberately a
bit hacky).

The only change I made to the table was to adjust the name to
`transfer_topportunity` - DMS lets you add a prefix to all tables in a
single rule, and it helps separate it a bit better from our existing
tables. This also occurs in the class name as `TransferTopportunity`
(although if we want to remove that `T` that would be fine) - just to
avoid a name clash with the `Opportunity` table and make clear it is
separate.

While it is possible to make the boolean and enum conversions something
that SQLAlchemy does on the serialize/deserialize logic of a column,
since this is the only table we intend to directly connect to the API, I
preferred a more direct implementation for now.

## Additional information
Local seed script works to populate the new transfer table.

<img width="999" alt="Screenshot 2024-01-29 at 12 08 09 PM"
src="https://github.com/HHS/simpler-grants-gov/assets/46358556/70e70f8e-f076-4b41-8722-7d80eaadffa4">

Local endpoint returns results uneventfully:
<img width="657" alt="Screenshot 2024-01-29 at 12 08 39 PM"
src="https://github.com/HHS/simpler-grants-gov/assets/46358556/aa040a49-4834-445f-b4e5-c1b31baa82c8">

All tests still working after adjusting some names for the factory
usage.

---------

Co-authored-by: nava-platform-bot <platform-admins@navapbc.com>