[Issue 1119] Deploy DMS from database module, fixup (#1129)
## Summary

Rolls up to #1119

### Time to review: __2 mins__

## Context for reviewers

I tried to deploy the DMS, and noticed 2 things:

- It didn't have a configuration for deploying it, like a file to run
`terraform init -backend-config=SOMETHING` on. So I was unable to pull
the existing state. This is why I moved the IAM role to the database
module, which is deployed from `infra/api/database`
- I got an error when deploying the database module. Specifically:
_"dms-access-role should have the DMS Regional Service Principal
'dms.us-east-1.amazonaws.com' as trusted entity"_

I then proceeded to deploy to `dev` and fix everything I could find

## Deployment

I've already deployed this to dev to test. It still doesn't fully deploy
- for reasons I will solve in a different PR tomorrow.