[Issue 1323] Fix a mis-firing alarm  (#1324)
## Summary 

Fixes #1323

### Time to review: __1 mins__

## Changes proposed

Makes the `high-app-response-time` alarm treat missing data as not
breaching.

## Context for reviewers

Right now this alarm is blowing up our inboxes because its swapping back
and forth between `INSUFFICIENT_DATA` and `OK`. It's doing this because
there's truly not enough data for most of the evaluation periods. My
suggested fix for this is to treat missing data as a non-issue, so we
don't get alerts when there is simply low traffic on the API.

## Testing

This is a "plan and pray" situation, I haven't tested this.