[Issue #1280] Remove DMS components (#1351)
## Summary
Fixes #1280

### Time to review: __5 mins__

## Changes proposed

Removes DMS components, except for the networking parts (eg. the peering
connection).

## Testing

I deployed this PR to `dev` to test

## Quirks

- This PR would normally delete a KMS key. Except, deleting KMS keys
violates an AWS security best practice. So instead I have changed the
description on the KMS key, and moved it slightly.
- This PR deletes a security group (inside
`infra/modules/dms-networking/security.tf`) that is being used by a
module (inside `infra/modules/database/networking.tf`). That forces an
order of operations on the deployment. Specifically, we must deploy in
this order:
1. deploy `infra/api/database/` to remove the use of the security group
  2. deploy `infra/networks/` to delete the security group

---------

Co-authored-by: Alsia Plybeah <alsia@navapbc.com>