Updated architecture diagram (#1499)
## Summary

Fixes https://github.com/HHS/simpler-grants-gov/issues/1276

### Time to review: __5 mins__

## Changes proposed

- removes the mermaid AWS architecture diagrams
- adds a lucidchart AWS architecture diagram

## Additional information

This architecture diagram is meant to be viewed as an abstraction of the
important high level components of our architecture. As such, there's
some fine details that are intentionally missing from it. For example,
the "VPC Endpoints" section intentionally doesn't list every single AWS
service we are accessing through VPC endpoints.

Given that, I'm interested in feedback on whether or not the
architecture diagram is missing some high level components. For example,
in early versions of this diagram I forgot the foreign data wrapper +
step functions setup. Large components like that are important to have
called out in the diagram, so please let me know if any are missing!

~Also note that this diagram pre-emptively includes New Relic.~