[Issue #1519] Cherry pick platform's pattern for env vars & ssm secrets (#1516)
## Summary

Relates to #784

Closes https://github.com/HHS/simpler-grants-gov/issues/1519

Copies https://github.com/navapbc/template-infra/pull/549

### Time to review: __10 mins__

## Changes proposed

- cherry picks platform infra template's pattern for passing in env vars
and AWS SSM secrets
- changes `ENABLE_V_0_1_ENDPOINTS` / `enable_v01_endpoints` to use the
above pattern
- _**does not yet**_ change any of our SSM secrets to use platform's
pattern, I plan to do that in a follow-up PR

## Context for reviewers

I created this PR via tactical copy-pasting from the
https://github.com/navapbc/template-infra/ repo.

The goal of this PR is to DRY our methods for setting environment
variables. Notice on the red side of the diff, how I've removed the need
to set `enable_v01_endpoints` so many times. Then notice on the green
side of the diff, that I only need to set `ENABLE_V_0_1_ENDPOINTS` twice
(for dev and staging). That's the goal of this PR, to pull in platform's
very nice pattern for DRY'ing environment variables.

## Testing

To test this, I added - then removed - the following block from
`staging.tf`

```hcl
  service_override_extra_environment_variables = {
    ENABLE_V_0_1_ENDPOINTS = "true"
  }
```

I then deployed to staging to see the difference. It worked as intended.