[Issue 1541] Ran `make lint` on analytics folder (#1539)
## Summary

Relates to https://github.com/HHS/simpler-grants-gov/issues/1541

### Time to review: __2 mins__

## Changes proposed

I ran `make lint` then committed most of the results. I didn't commit
all of the results because this isn't my code and so I don't want to
overstep. Which is to say, `make lint` produced even more changes that
you don't see in this PR. I think it would ideally be the responsibility
of the code author(s) to address those changes.

## Context for reviewers

This PR is necessary due to CI not actually checking if the lining is
being done. See https://github.com/HHS/simpler-grants-gov/issues/1541
for more details.