[Issue #1521]: fix API_AUTH_TOKEN clash between backend and frontend (#1759)
## Summary

Fixes #1521

### Time to review: __2 mins__

## Context

Both `infra/api` and `infra/frontend` were defining a variable called
`API_AUTH_TOKEN`. The way they were defined made them clash, leading to
the following error message:

> Error: creating ECS Task Definition (frontend-dev): ClientException:
The secret name must be unique and not shared with any new or existing
environment variables set on the container, such as 'API_AUTH_TOKEN'

## Changes proposed

Refactors the terraform configuration so the `API_AUTH_TOKEN` definition
is no longer clashing between the frontend and backend

## Testing

I have already deployed both the API and the frontend to dev to test
that this is working.