[Issue #1659] Create general purpose s3 bucket (#1762)
## Summary

Fixes #1659

### Time to review: __4 mins__

## Changes proposed

Adds a "General Purpose" s3 bucket to all of our services. The s3 bucket
exposes its location via the `S3_BUCKET_ARN` environment variable. I
assume this is going to be a long lived general purpose s3 bucket for...
random stuff. Whatever you might want an S3 bucket for. It's also
specifically for populating seed data, eg.
https://github.com/HHS/simpler-grants-gov/issues/1659

## Context for reviewers

The S3 bucket itself is a near exact copy of the s3 bucket that we use
for access logs.

## Additional information

Note that the s3 bucket arn format is like so:
`arn:aws:s3:::api-dev-general-purpose20240416221336163000000001`. You
probably want to `.split(":")[-1]` it to get just the name.