[Issue #1736] Adjust current opportunity summary logic to want null revision number (#1743)
## Summary
Fixes #1736

### Time to review: __3 mins__

## Changes proposed
Modify the set-current-opportunity script to determine the latest
opportunity summary based on the revision number being null (rather than
the max value)

## Context for reviewers
My understanding of the revision number in the prior implementation was
incorrect. We want to determine the latest revision of an opportunity
summary based on the revision number, however the latest in the current
Oracle system will always be null. The way the current system works is
that the revision number is only added once it is moved to the history
table. Because we effectively want to ignore all of the historical
synopsis/forecast records, we can just check if the value is null.

Due to how the data is structured in the current system where the values
are unique in the synopsis & forecast tables for a given opportunity, we
know there will only ever be a single forecast / non-forecast that has
this value set to null in the data.