[Issue #1659] Write a script that copies csvs from s3 to the db + script to generate them (#1773)
## Summary
Fixes #1659

### Time to review: __5 mins__

## Changes proposed
**this is temporary code and will be deleted after we run it - these are
onetime scripts and are deliberately a bit hacky**

Adds two scripts:
* One that takes csvs extracts of the Oracle database tables and does
transformations to make csvs we can import into our database
* One that takes the output of the first script and uploads them into
our database

## Context for reviewers
As the transformation work is just starting, we want to unblock testing
of the front-end UI by getting data into our dev environment. This
approach is basically a hacky one-time shortcut to get data uploaded so
that we have time to build the actual transformation process.
Additionally, we using prod data so that it's realistic for testing
purposes.

The unit test I setup for the load script just exists to verify the
tools work if the input dataset is on disk (as I did locally) or via S3
(as it will actually be run).

## Additional information
Testing this requires downloading data from the real production
database, which I think only I have access to at the moment. I did
actually build this out / test it using that data locally. You can also
use the test files I uploaded for the unit test to verify the behavior
of the second part of the script.


To run it:
```
poetry run python tests/util/convert_oracle_csvs_to_postgres.py --directory <whatever directory your csvs are in>

poetry run flask task import-opportunity-csvs --input-folder <same directory as before>
```

Note this will not work at the moment due to needing a change from
https://github.com/HHS/simpler-grants-gov/pull/1743