[Issue #1835] Remove distinct from inner query of search (#1836)
## Summary
Fixes #1835

### Time to review: __3 mins__

## Changes proposed
Adjust the search query to not do `distinct` on the inner filter query

## Context for reviewers

This distinct originally was added as the query was setup in a much
different way. We needed to do the distinct to avoid duplicates when
doing joins, but because it is now an inner query - quite literally used
as `select ... where opportunity_id in (select distinct opportunity_id
...)` we don't need to distinct it. If the values aren't distinct, it
doesn't actually change the outer query, there will just be dupes in
that list, but that's fine and seems to be better performing.