[Issue #1853] Move where we define the lookup mapping to avoid constant import issues (#1854)
## Summary
Fixes #1853

### Time to review: __1 mins__

## Changes proposed
Move where we define the LookupConfig definitions to reduce circular
dependencies

## Context for reviewers
This doesn't change the configs, just where they're defined. We
frequently want to import the enums defined in lookup_constants, but
because that file in turn depends on the DB lookup files, we run into
some odd circular dependencies that require manually sorting imports to
resolve. Instead, we'll just define those LookupConfig classes in the
only place they're ever used to avoid the headache.