[Issue #1905] Add unique constraint for opportunity summary (#1906)
## Summary
Fixes #1905

### Time to review: __5 mins__

## Changes proposed
Upgrade our local version of Postgres to version 15

Add a unique constraint to the opportunity summary for

## Context for reviewers
This constraint is something that is currently implicitly handled in the
legacy system across 4 tables, but we're merging into one table. Adding
this constraint makes us more certain that we're not importing data in
the transformation process incorrectly.

- `is_forecast` - whether it came from the forecast or synopsis table
(or their history equivalents)
- `revision_number` - null from the non-historical tables, otherwise a
part of the primary key
- `opportunity_id` - part of the primary key in all of the source Oracle
tables

Together they let us uniquely point back to a record in the legacy
system without issue.

Note that because revision number can be null, to make it work correctly
in a unique constraint we need to use the `NULLS NOT DISTINCT` flag when
creating the unique index:
https://www.postgresql.org/docs/current/indexes-unique.html - this was
only added in Postgres 15. While there are some other workarounds in
older versions, we're running Postgres 15 in AWS, so I see zero reason
not to upgrade and match. What this flag does is basically make it so in
a compound unique key, null is treated as a distinct value. For example,
lets say we have values of `True, null, 1`. Without `NULLS NOT DISTINCT`
you could keep inserting this same 3 values because it would think null
!= null.

## Additional information
Note that you will need to run `make db-recreate` as the volume that
backs a v14 and v15 Postgres DB are incompatible. This will drop your DB
and remake it.